data_api/requirements.txt
sqlalchemy
sqlalchemy[mypy]
flask
flask-sqlalchemy
GitHub-Flask
google-cloud-secret-manager
cloud-sql-python-connector[pg8000]
